requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pyyaml==6.0.2
watchdog==6.0.0
